Song

+++ app/src/main/java/com/example/front_ui/DataModel/StoreInfo.java
@@ -1,16 +1,20 @@
 package com.example.front_ui.DataModel;
 
+import android.os.Parcel;
+import android.os.Parcelable;
+
 import com.google.firebase.firestore.GeoPoint;
 
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
 public class StoreInfo{
     public String name;
     public String storeId;
-    public float aver_star;
+    public double aver_star;
     public String address;
-    public List<Float> detail_aver_star;
+    public List<Double> detail_aver_star;
     private ArrayList<PostingInfo> allItemsInSection;
     public GeoPoint geoPoint;
 
@@ -22,7 +26,7 @@ public class StoreInfo{
     public StoreInfo(String name,
                      float aver_star,
                      String address,
-                     List<Float> detail_aver_star,
+                     List<Double> detail_aver_star,
                      GeoPoint geoPoint
                      ) {
         this.name = name;
@@ -39,7 +43,7 @@ public class StoreInfo{
         return name;
     }
 
-    public float getAver_star() {
+    public double getAver_star() {
         return aver_star;
     }
 
@@ -47,7 +51,7 @@ public class StoreInfo{
         return address;
     }
 
-    public List<Float> getDetail_aver_star() {
+    public List<Double> getDetail_aver_star() {
         return detail_aver_star;
     }
 
@@ -81,4 +85,27 @@ public class StoreInfo{
     public void setStoreId(String storeId) {
         this.storeId = storeId;
     }
+
+//    @Override
+//    public int describeContents() {
+//        return 0;
+//    }
+//
+//    @Override
+//    public void writeToParcel(Parcel dest, int i) {
+//        dest.writeString(name);
+//        dest.writeString(storeId);
+//        dest.writeFloat(aver_star);
+//        dest.writeString(address);
+//        dest.writeO(detail_aver_star);
+//        dest.writeTypedObject(geoPoint);
+//
+//        public String name;
+//        public String storeId;
+//        public float aver_star;
+//        public String address;
+//        public List<Float> detail_aver_star;
+//        private ArrayList<PostingInfo> allItemsInSection;
+//        public GeoPoint geoPoint;
+//    }
 }
